(PC-33377)[PRO] refactor: Removes 'pages/VenueCreation' folder and moves shared components & utils where they should belong
