+++ pro/src/commons/core/shared/utils/validation.ts
@@ -4,9 +4,6 @@ import { SelectOption } from 'commons/custom_types/form'
 
 import { parseAndValidateFrenchPhoneNumber } from './parseAndValidateFrenchPhoneNumber'
 
-export const isOneTrue = (values: Record<string, boolean>): boolean =>
-  Object.values(values).includes(true)
-
 export const isPhoneValid = (phone: string | undefined): boolean => {
   if (!phone) {
     return false
