+++ pro/src/components/SignupJourneyForm/Activity/ActivityForm.tsx
@@ -5,7 +5,7 @@ import { VenueTypeResponseModel } from 'apiClient/v1'
 import { FormLayout } from 'components/FormLayout/FormLayout'
 import fullMoreIcon from 'icons/full-more.svg'
 import fullTrashIcon from 'icons/full-trash.svg'
-import { buildVenueTypesOptions } from 'pages/VenueCreation/buildVenueTypesOptions'
+import { buildVenueTypesOptions } from 'pages/VenueEdition/buildVenueTypesOptions'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { CheckboxGroup } from 'ui-kit/form/CheckboxGroup/CheckboxGroup'
