--- pro/src/pages/VenueCreation/VenueCreationForm.module.scss
@@ -1,5 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-
-.eac-description-info {
-  margin: rem.torem(16px) 0 rem.torem(24px) 0;
-}
