--- pro/src/pages/VenueCreation/VenueCreationFormScreen.module.scss
@@ -1,33 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-
-.venue-form-heading {
-  margin-bottom: rem.torem(32px);
-}
-
-.title {
-  @include fonts.title1;
-}
-
-.title-page {
-  display: flex;
-  justify-content: space-between;
-}
-
-.venue-name {
-  margin-bottom: rem.torem(8px);
-}
-
-.identifier {
-  &-hidden {
-    position: absolute;
-    width: 0;
-    height: 0;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip: rect(0, 0, 0, 0);
-    white-space: nowrap;
-    border: 0;
-  }
-}
