+++ pro/src/pages/VenueEdition/constants.ts
@@ -1,0 +1,7 @@
+export const DEFAULT_INTITIAL_OPENING_HOURS = {
+  morningStartingHour: '',
+  morningEndingHour: '',
+  afternoonStartingHour: '',
+  afternoonEndingHour: '',
+  isAfternoonOpen: false,
+}
